requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
streamlit==1.35.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pyrebase4==4.7.1
firebase-admin==6.5.0
python-multipart==0.0.9
requests==2.31.0
pandas==2.2.2
Pillow==10.3.0
httpx==0.27.0
